requirements.txt
anthropic>=0.3.0
pandas>=1.5.0
python-dotenv>=0.19.0